Consolidate duplicate sync logic to hook
