+++ projects/packages/forms/src/blocks/field-textarea/edit.js
@@ -1,16 +1,12 @@
-import {
-	useBlockProps,
-	useInnerBlocksProps,
-	store as blockEditorStore,
-} from '@wordpress/block-editor';
-import { useSelect, useDispatch } from '@wordpress/data';
-import { useMemo, useEffect } from '@wordpress/element';
+import { useBlockProps, useInnerBlocksProps } from '@wordpress/block-editor';
+import { useMemo } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import clsx from 'clsx';
 import JetpackFieldControls from '../shared/components/jetpack-field-controls';
 import useFieldSelected from '../shared/hooks/use-field-selected';
 import useFormWrapper from '../shared/hooks/use-form-wrapper';
 import useJetpackFieldStyles from '../shared/hooks/use-jetpack-field-styles';
+import useSyncRequiredIndicator from '../shared/hooks/use-sync-required-indicator';
 import { ALLOWED_INNER_BLOCKS } from '../shared/util/constants';
 
 export default function TextareaFieldEdit( props ) {
@@ -41,27 +37,7 @@ export default function TextareaFieldEdit( props ) {
 		templateLock: 'all',
 	} );
 
-	// Keep the inner label block's requiredIndicator in sync when it changes
-	const labelClientId = useSelect(
-		select => {
-			const { getBlock } = select( blockEditorStore );
-			const parentBlock = getBlock( clientId );
-			if ( ! parentBlock ) {
-				return undefined;
-			}
-			const labelBlock = parentBlock.innerBlocks.find( block => block.name === 'jetpack/label' );
-			return labelBlock?.clientId;
-		},
-		[ clientId ]
-	);
-
-	const { updateBlockAttributes } = useDispatch( blockEditorStore );
-
-	useEffect( () => {
-		if ( labelClientId ) {
-			updateBlockAttributes( labelClientId, { requiredIndicator } );
-		}
-	}, [ labelClientId, requiredIndicator, updateBlockAttributes ] );
+	useSyncRequiredIndicator( clientId, requiredIndicator );
 
 	return (
 		<>
